ai-agents/requirements.txt
PyYAML>=6.0
aiohttp>=3.8.0
requests>=2.28.0